Cleaned Up Testing Files
