--- SW5E/Assets/Scripts/Backend.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f0bbf6be02dd4f944a5b6b3d50f183e6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
